all but completed the exercise!!
this should have everything in it except for a functioning goal and a reset system when the player dies (everything that needs to reset, save for a few pickups that boost score, already has an event. i just need to write the system that triggers it.)

i'm so sorry! i know i said i'd have this in on tuesday, but i've been working all night and i have class in the morning. as soon as i can (and after i make up for lost sleep), the rest of the exercise will be put in